add IndexerFileManager
